Moves the redis key update so that it gets set on each iteration.
This is an attempt to cure the issue where some messages get 'stuck'
and get sent repeatedly to the HipChat room. I think this may be be-
cause there is an exception which prevents the timestamp from being
updated so that the dodgy message is repeatedly fetched. This way,
even if a message borks, so long as we get the timestamp we should
be able to prevent it from repeating on us. Like a kipper.